backend/requirements.txt
flask
flask-cors
torch
torchvision
opencv-python
mediapipe
pillow
numpy
sqlite3-binary
